pyproject.toml
[project]
name = "video-summarizer"
version = "0.1.0"
description = "A CLI video summarizer tool that helps you pull videos off youtube and creates a markdown summary file with important keyframes"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "Video Summarizer Team"}
]
license = {text = "MIT"}
keywords = ["video", "summarizer", "youtube", "cli", "keyframes", "transcription"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "click>=8.1.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "yt-dlp>=2023.12.0",
    "opencv-python>=4.8.0",
    "pillow>=10.0.0",
    "numpy>=1.24.0",
    "requests>=2.31.0",
    "pathlib",
    "imagehash>=4.3.0",
    "tqdm>=4.65.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
]

[project.scripts]
video-summarizer = "video_summarizer.main:main"
vs = "video_summarizer.main:main"

[project.urls]
Homepage = "https://github.com/yourusername/video-summarizer"
Repository = "https://github.com/yourusername/video-summarizer"
Issues = "https://github.com/yourusername/video-summarizer/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
